chore: move link components into single file
