--- src/components/links/external-link.tsx
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const ExternalLink = styled.a`
-  text-decoration: none;
-  color: inherit;
-`;
